Header for mobile settings menu
